% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descarga_datos.R
\name{descarga_data}
\alias{descarga_data}
\title{Descargar y leer datos de estaciones meteorológicas}
\usage{
descarga_data(id_estacion, ruta)
}
\arguments{
\item{id_estacion}{Código de la estaciónes disponibles: "NH0437,"NH0472", "NH0046", "NH0910", "NH0098".}

\item{ruta}{Ruta donde se guardará el archivo descargado, por ejemplo "datos/NH0472.csv"}
}
\value{
Un data frame con los datos de la estación
}
\description{
Esta función descarga (si es necesario) y lee los datos de distintas estaciones meteorológicas
a partir de su ID.
}
\examples{
\dontrun{
datos <- descarga_data("NH0472", "datos/NH0472.csv")
}
}
