% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resumen_lluvias.R
\name{resumen_lluvias}
\alias{resumen_lluvias}
\title{Resumen de precipitaciones por mes}
\usage{
resumen_lluvias(datos)
}
\arguments{
\item{datos}{Un data frame que contenga al menos las columnas 'id', 'fecha' y 'precipitacion_pluviometrica'.}
}
\value{
Un data frame con las columnas:
\itemize{
\item id - identificador de la estación.
\item fecha
\item lluvia_total - cantidad total en el mes.
\item lluvia_promedio - promedio diario a lo largo del mes.
\item lluvia_ocurrencia - cantidad de dias que llovio.}
}
\description{
Calcula estadisticas basicas de precipitacion a partir de un data set con las variables requeridas.
Devuelve un resumen con el total de lluvia diaria,
cantidad de días con precipitación y promedio por estación.
}
\examples{
\dontrun{
resumen_lluvias(datos)
}


}
