% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabla_resume_temperatura.R
\name{tabla_resume_temperatura}
\alias{tabla_resume_temperatura}
\title{Tabla resumidora de temperatura a 150 cm}
\usage{
tabla_resume_temperatura(datos)
}
\arguments{
\item{datos}{Un data frame que contenga al menos las columnas 'id' y 'temperatura_abrigo_150cm'}
}
\value{
Un data frame con las columnas:
\itemize{
\item id - identificador de la estación.
\item promedio - temperatura promedio a 150 cm.
\item mediana - temperatura mediana a 150 cm.
\item minimo - temperatura mínima a 150 cm.
\item maximo - temperatura máxima a 150 cm. }
}
\description{
Genera una tabla de resumen estadístico de la variable 'temperatura_abrigo_150cm'
para una o más estaciones meteorológicas.
}
\examples{
\dontrun{
datos<- descarga_data("NH0472", "datosmeteo/NH0472.csv")
tabla_resume_temperatura(datos)
}


}
