% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualtm.R
\name{visualtm}
\alias{visualtm}
\title{Gráfico de temperatura mensual por estación}
\usage{
visualtm(datos, colores = NULL, titulo = "Temperatura")
}
\arguments{
\item{datos}{Un data frame con las columnas 'id', 'fecha', y 'temperatura_abrigo_150cm'.}

\item{colores}{Vector opcional de colores personalizados para las estaciones.}

\item{titulo}{Título del gráfico. Por defecto, "Temperatura".}
}
\value{
Un objeto ggplot2 con el gráfico de temperaturas mensuales.
}
\description{
Esta función genera un gráfico de líneas de la temperatura promedio mensual
a 150 cm de abrigo. Si se desea graficar más de una estación, se pueden
combinar los datasets usando \code{dplyr::bind_rows()} antes de llamar a esta función.
}
\examples{
\dontrun{
datos <- descarga_data("NH0472", "datosmeteo/NH0472.csv")
visualtm(datos)
}
}
